--- dev/e2e/README.md
@@ -1,1 +1,0 @@
-# Flutter end-to-end driver tests
