+++ dev/end_to_end/README.md
@@ -1,0 +1,3 @@
+# Flutter end-to-end driver tests
+
+This is a collection of end-to-end integration tests for Flutter.
