+++ dev/end_to_end/android/app/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
      found in the LICENSE file.
  -->
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.yourcompany.e2e"
+    package="com.yourcompany.end_to_end"
     android:versionCode="1"
     android:versionName="0.0.1">
 
@@ -20,7 +20,7 @@
          In most cases you can leave this as-is, but you if you want to provide
          additional functionality it is fine to subclass or reimplement
          FlutterApplication and put your custom class here. -->
-    <application android:name="io.flutter.app.FlutterApplication" android:label="e2e">
+    <application android:name="io.flutter.app.FlutterApplication" android:label="EndToEnd">
         <activity android:name=".MainActivity"
                   android:launchMode="singleTop"
                   android:theme="@android:style/Theme.Black.NoTitleBar"
