+++ dev/end_to_end/android/app/src/main/java/com/yourcompany/end_to_end/MainActivity.java
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package com.yourcompany.e2e;
+package com.yourcompany.end_to_end;
 
 import android.os.Bundle;
 import io.flutter.app.FlutterActivity;
