+++ dev/end_to_end/android/gradle.properties
