+++ dev/end_to_end/android/settings.gradle
