+++ dev/end_to_end/test_driver/main_test.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:e2e/keys.dart' as keys;
+import 'package:end_to_end/keys.dart' as keys;
 import 'package:flutter_driver/flutter_driver.dart';
 import 'package:test/test.dart';
